Feat(queue.c/.h): Implement q_size
The implementation of q_size features:
* using size_t for return type
* if q is NULL return 0

+++ queue.h
@@ -75,7 +75,7 @@ bool q_remove_head(queue_t *q, char *sp, size_t bufsize);
  * Return number of elements in queue.
  * Return 0 if q is NULL or empty
  */
-int q_size(queue_t *q);
+size_t q_size(queue_t *q);
 
 /*
  * Reverse elements in queue
